requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cycler==0.10.0
distlib==0.3.1
Django==3.1.3
django-bootstrap4==2.3.1
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
html5lib==1.1
idna==2.10
imaplib2==2.45.0
keyring==21.8.0
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.4
matplotlib==3.3.3
msgpack==1.0.2
numpy==1.19.3
opencv-python==4.1.2.30
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.0.1
pkginfo==1.7.0
poetry==1.1.6
poetry-core==1.0.3
progressbar==2.5
psycopg2==2.8.6
ptyprocess==0.7.0
pylev==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.25.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tomlkit==0.7.0
urllib3==1.26.4
virtualenv==20.4.3
webencodings==0.5.1
